requirements.txt
Django>=2.0
djangorestframework>=3.8.2
django-uuslug>=1.1.8
dj-database-url>=0.5.0
gunicorn>=19.8.1
psycopg2
whitenoise>=3.3.1
